handle detection fields with heterogenous tracks/shapes

+++ fiftyone/utils/cvat.py
@@ -2664,7 +2664,8 @@ class CVATAnnotationResults(foua.AnnotationResults):
         """
         api = self.connect_to_api()
         api.delete_tasks(self.task_ids)
-        api.delete_projects(self.project_ids)
+        if self.project_ids:
+            api.delete_projects(self.project_ids)
 
         # @todo save updated results to DB?
         self.project_ids = []
@@ -5552,7 +5553,8 @@ def _get_interpolated_shapes(track_shapes):
     prev_shape = {}
     for shape in track_shapes:
         if prev_shape:
-            assert shape["frame"] > curr_frame
+            if shape["frame"] <= curr_frame:
+                continue
             for attr in prev_shape["attributes"]:
                 if attr["spec_id"] not in map(
                     lambda el: el["spec_id"], shape["attributes"]
